Separate Env-related stuff from GHC.hs
Move this functionality to a new module: Distribution.Gentoo.Env

Rework the EnvM monad into a more generic EnvT monad transformer with
EnvIO, EnvWriterT, and EnvId variants. Add a new class, MonadEnv, to
increase generalization and help facilitate newtypes over EnvT.

Replace runEnvM with collectEnv to collect system state, which can then
be passed to runEnvT.

SayM has been reworked into a SayT monad transformer.
