+++ haskell-updater.cabal
@@ -65,6 +65,7 @@ library haskell-updater-internal
     , process
 
   exposed-modules:
+    Distribution.Gentoo.Env
     Distribution.Gentoo.GHC
     Distribution.Gentoo.Packages
     Distribution.Gentoo.PkgManager
